requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
psycopg2-binary>=2.9.5
pgvector>=0.2.0
nltk>=3.8.1
spacy>=3.5.0
reportlab>=3.6.12
tqdm>=4.65.0 